+++ rust-src/concordium_base/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Improve performance of `multiexp*` family of functions.
 - Support `P7` protocol version.
+- The `Debug` implementation for `ContractEvent` displays the value in `hex`.
+  The alternate formatter (using `#`) displays it as a list of bytes.
 
 ## 3.2.0 (2023-11-22)